requirements.txt
streamlit==1.30.0
plotly==5.9.0
numpy==1.24.3
pandas==1.5.3
folium==0.14.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
streamlit-folium==0.13.0
Pillow==9.4.0
image==1.5.33
XlsxWriter==3.1.9